исправляет ошибки, исправляет вывод городов при сортировке
